fix(hipcub): use int64_t instead of long long

+++ projects/hipcub/benchmark/benchmark_block_discontinuity.cpp
@@ -240,8 +240,8 @@ void add_benchmarks(primbench::executor& executor)
     BENCHMARK_TYPE(int8_t, 256, true);
     BENCHMARK_TYPE(uint8_t, 256, false);
     BENCHMARK_TYPE(uint8_t, 256, true);
-    BENCHMARK_TYPE(long long, 256, false);
-    BENCHMARK_TYPE(long long, 256, true);
+    BENCHMARK_TYPE(int64_t, 256, false);
+    BENCHMARK_TYPE(int64_t, 256, true);
 }
 
 int main(int argc, char* argv[])
